ip-reputation/requirements.txt
dnspython==2.6.1
python-json-logger==2.0.7